chore: render.js 에서 필터링 로직 분리
