+++ week2/assign/util/render.js
@@ -1,5 +1,6 @@
 import { members } from "../data.js";
 import { getFilteringValue } from "./getFilteringValue.js";
+import { filterMembers } from "./filterMembers.js";
 
 // 멤버 로컬스토리지에 저장
 const MEMBER_KEY = "membersData";
@@ -55,25 +56,8 @@ handleSearch.addEventListener("click", (event) => {
     week1: getFilteringValue("week1Input"),
     week2: getFilteringValue("week2Input"),
   };
-
-  // 공백, 대소문자 구분 없이 필터링
-
-  console.log(filters); // filters 객체 확인
-  console.log(memberList); // 원본 멤버 리스트 확인
-  const filteredMembers = memberList.filter((member) =>
-    Object.entries(filters).every(([key, value]) => {
-      const trimmedValue = value.trim().toLowerCase();
-      if (key === "gender") {
-        return (
-          trimmedValue === "" || member[key]?.toLowerCase() === trimmedValue
-        );
-      }
-      return (
-        trimmedValue === "" ||
-        member[key]?.toString().toLowerCase().includes(trimmedValue)
-      );
-    })
-  );
+  // 렌더링 할 최종 필터링된 멤버
+  const filteredMembers = filterMembers(memberList, filters);
 
   renderMemberList(filteredMembers);
 });
